esercizi-openai/pyproject.toml
[project]
name = "esercizi-openai"
version = "0.1.0"
description = "Esercizi OpenAI API - Corso Aulab"
authors = [
    {name = "Francesco Scaccetti",email = "scaccettif@gmail.com"}
]
requires-python = ">=3.13"
dependencies = [
    "openai (>=2.32.0,<3.0.0)",
    "python-dotenv (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
